backend/requirements.txt

snowflake-connector-python
python-dotenv
openai
pandas
google-api-python-client